+++ cumplo_herald/__init__.py
@@ -1,0 +1,3 @@
+"""A multi-channel notification service designed to keep Cumplo investors ahead with real-time alerts"""
+
+__version__ = "0.1.1"
