+++ cumplo_herald/business/channels/__init__.py
@@ -4,7 +4,7 @@ from logging import getLogger
 
 from cumplo_common.models.user import User
 
-from models.channel import Channel
+from cumplo_herald.models.channel import Channel
 
 logger = getLogger(__name__)
 
@@ -14,6 +14,9 @@ def import_channels(user: User) -> Iterator[Channel]:
     Imports the channels for the given user.
     """
     for channel_type, channel_configuration in user.channels.items():
+        if not channel_configuration.enabled:
+            logger.info(f"Channel {channel_type} is disabled")
+            continue
         try:
             module = import_module(f"business.channels.{channel_type}")
             channel = getattr(module, "channel")
