+++ cumplo_herald/business/writers/funding_requests.py
@@ -1,12 +1,14 @@
+# mypy: disable-error-code="override"
+
 from decimal import Decimal
 
 from cumplo_common.models.channel import ChannelType
 
-from business.channels.ifttt import IFTTTMessage
-from models.channel import Channel
-from models.message import Message
-from models.writer import Writer
-from schemas.topics.funding_requests import FundingRequestContent
+from cumplo_herald.business.channels.ifttt import IFTTTMessage
+from cumplo_herald.models.channel import Channel
+from cumplo_herald.models.message import Message
+from cumplo_herald.models.topics.funding_requests import FundingRequestContent
+from cumplo_herald.models.writer import Writer
 
 
 class PromisingFundingRequestsWriter(Writer):
